chore: metaservice addon moved to proxy-core
Refs: XRDDEV-2829

+++ src/service/proxy/proxy-core/src/test/java/org/niis/xroad/proxy/core/testsuite/ProxyTests.java
@@ -210,7 +210,9 @@ public class ProxyTests {
 
     @AfterEach
     void afterEach() {
-        ctx.destroy();
+        if (ctx != null) {
+            ctx.destroy();
+        }
         ctx = null;
     }
